[timing] only output whole numbers to CSV file
Context: https://jenkins.mono-project.com/view/Xamarin.Android/job/xamarin-android/plot/Build%20times/

The Jenkins plot plugin does not seem to be able to render decimal
numbers in the table displaying the CSV file.

Radek has sent a PR upstream to the plugin: https://github.com/jenkinsci/plot-plugin/pull/47

In the meantime, we can fix this ourselves by only emitting whole
numbers to the CSV file. Since the build times are taking multiple
seconds and we are storing milliseconds in the CSV, this loss of
precision won't affect us much.

+++ build-tools/Xamarin.Android.Tools.BootstrapTasks/Xamarin.Android.Tools.BootstrapTasks/ProcessMSBuildTiming.cs
@@ -29,7 +29,7 @@ namespace Xamarin.Android.Tools.BootstrapTasks
 				var id = build.Attribute ("id")?.Value;
 				var elapsed = build.Attribute ("elapsed")?.Value;
 				if (TimeSpan.TryParse (elapsed, out TimeSpan result)) {
-					results [id] = result.TotalMilliseconds.ToString ();
+					results [id] = ((int)result.TotalMilliseconds).ToString ();
 				}
 			}
 			WriteResults ();
